Split input-handling into a separate script
