+++ Assets/Fungus/Scripts/Components/MenuDialog.cs
@@ -8,6 +8,7 @@ using UnityEngine.EventSystems;
 using System.Linq;
 using MoonSharp.Interpreter;
 using Fungus.Lua;
+using System;
 
 #if ENABLE_INPUT_SYSTEM
 using UnityEngine.InputSystem;
@@ -22,20 +23,6 @@ namespace Fungus
 	{
 		[Tooltip("Automatically select the first interactable button when the menu is shown.")]
 		[SerializeField] protected bool autoSelectFirstButton = false;
-
-		[Header("Input-Handling")]
-		[Tooltip("Automatically select the first interactable button when one of the inputs are triggered" +
-			"and none of the buttons were selected at the time. This is mainly intended to make it so this" +
-			"responds well to transitioning between mouse/keyboard and gamepad.")]
-		[SerializeField] protected bool autoSelectBasedOnInput = true;
-#if ENABLE_LEGACY_INPUT_MANAGER
-		[Tooltip("In response to any of these axes, this will make sure there's one button selected in the menu dialog.")]
-		[SerializeField] protected string[] inputAxes = new string[0];
-#endif
-#if ENABLE_INPUT_SYSTEM
-		[Tooltip("In response to any of these actions, this will make sure there's one button selected in the menu dialog.")]
-		[SerializeField] protected InputActionReference[] inputActions = new InputActionReference[0];
-#endif
 
 		protected virtual void Awake()
 		{
@@ -140,85 +127,8 @@ namespace Fungus
 			// The canvas may fail to update if the menu dialog is enabled in the first game frame.
 			// To fix this we just need to force a canvas update when the object is enabled.
 			Canvas.ForceUpdateCanvases();
-#if ENABLE_INPUT_SYSTEM
-			EnableAndListenForInput();
-
-			void EnableAndListenForInput()
-			{
-				foreach (InputActionReference actionEl in inputActions)
-				{
-					if (actionEl != null && actionEl.action != null)
-					{
-						actionEl.action.Enable();
-						actionEl.action.performed += OnActionPerformed;
-					}
-				}
-			}
-#endif
 		}
 
-#if ENABLE_INPUT_SYSTEM
-		private void OnActionPerformed(InputAction.CallbackContext context)
-		{
-			EnsureOneOptionIsSelected();
-		}
-
-		protected virtual void EnsureOneOptionIsSelected()
-		{
-			bool anyOptionsSelected = CachedButtons.Any
-				(
-				option => option.gameObject.activeInHierarchy
-				&& option.interactable &&
-				EventSystem.current.currentSelectedGameObject == option.gameObject
-				);
-
-			if (!anyOptionsSelected)
-			{
-				Button toSelect = CachedButtons.FirstOrDefault(option => option.gameObject.activeInHierarchy && option.interactable);
-
-				if (toSelect != null)
-				{
-					EventSystem.current.SetSelectedGameObject(toSelect.gameObject);
-				}
-			}
-		}
-		protected virtual void OnDisable()
-		{
-			UNlistenForInput();
-			void UNlistenForInput()
-			{
-				foreach (var actionEl in inputActions)
-				{
-					if (actionEl != null && actionEl.action != null)
-					{
-						actionEl.action.performed -= OnActionPerformed;
-					}
-				}
-			}
-		}
-#endif
-
-		protected virtual void Update()
-		{
-#if ENABLE_LEGACY_INPUT_MANAGER
-			HandleResponseToInputAxes();
-#endif
-		}
-
-#if ENABLE_LEGACY_INPUT_MANAGER
-		protected virtual void HandleResponseToInputAxes()
-		{
-			foreach (string inputAxisEl in inputAxes)
-			{
-				bool inputDetected = Input.GetAxis(inputAxisEl) != 0;
-				if (inputDetected)
-				{
-					EnsureOneOptionIsSelected();
-					return; // So we don't iterate over more axes per frame than necessary
-				}
-			}
-		}
-#endif
 		#region Public members
 
 		/// <summary>
@@ -242,7 +152,6 @@ namespace Fungus
 		{
 			if (ActiveMenuDialog == null)
 			{
-
 				// Use first Menu Dialog found in the scene (if any)
 			#if UNITY_6000
 				var menuDialogFound = FindFirstObjectByType<MenuDialog>();
