+++ Assets/Fungus/Scripts/Input/MenuDialogInput.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 72a23fa110aefcf4da64ef5e32a83971
